LangChain_ai_chat_agent/requirements.txt
langchain>=0.3
langchain-core>=0.3
langchain-deepseek>=0.1
langchain-community>=0.3
tavily-python>=0.5
langgraph>=0.2
langgraph-checkpoint-mysql[pymysql]>=2.0
cryptography>=42.0.0
python-dotenv>=1.0
tzdata>=2024.1
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0
asyncmy>=0.2
PyJWT>=2.8
bcrypt>=4.1
